requirements.txt
quart==0.15.1
quart-cors==1.1.0
pandas==1.3.4
beautifulsoup4==4.10.0
openpyxl==3.0.9
numpy==1.21.3
langchain==0.1.0
PyPDF2==1.26.0
pinecone-client==0.9.0
